stair-climber/requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorzero==2.0
Django==3.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gpiozero==1.6.2
grpcio==1.34.1
h5py==3.1.0
idna==2.10
image==1.5.33
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyboard==0.13.5
lxml==4.6.3
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.4
pathfinding==1.0.1
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynput==1.7.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc1
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
